libstdc++: replace std::powf/sqrtf calls
Include cmath for using std::pow and std::sqrt, and avoid using f
variants as they are not available in libstdc++

Bug: 957519
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4964679
Commit-Queue: José Dapena Paz <jdapena@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1214045}

+++ media/gpu/exponential_moving_average.cc
@@ -5,6 +5,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "media/gpu/exponential_moving_average.h"
 
+#include <cmath>
+
 #include "base/check.h"
 #include "base/check_op.h"
 #include "base/logging.h"
@@ -18,7 +20,7 @@ ExponentialMovingAverage::ExponentialMovingAverage(
 ExponentialMovingAverage::~ExponentialMovingAverage() = default;
 
 float ExponentialMovingAverage::GetStdDeviation() const {
-  return std::sqrtf(std::max(mean_square_ - std::pow(mean_, 2), 0.0));
+  return std::sqrt(std::max(mean_square_ - std::pow(mean_, 2.0f), 0.0f));
 }
 
 }  // namespace media
